Fix windows builds, ifort and gfortran tested
Updated cmakelist.txt

Should only need remove /fpp now

Fix windows build

Some symbols may still need extra dllimport to work.

Minor fixes. Geant4Interface example should work now

Also, all build should be passing.

Do note that cmake gives /fpp CMAKE_Fortran_FLAGS on ifort. This breaks build with error:
```
```
Removing that option in cmake cache file or specifying that in compile options is required to build.

This may correspond to the `-std=legacy` option in original cmake file, but not supported on ifort.

Add comment on build on windows.

make options conditional, remove /fpp by using /nofpp

Add windows build instructions

Move ifort logic outside win32 platform

Add comment on 64 bit

clarify compiler settings

Now things should work no matter what

Add ignore binaries

+++ Examples/DriftTube/mdt_mt.C
@@ -129,7 +129,7 @@ int main(int argc, char * argv[]) {
     // Loop over the primary electrons along the track.
     const std::size_t ne = electrons.size();
     #pragma omp parallel for
-    for (size_t k = 0; k < ne; ++k) {
+    for (int k = 0; k < ne; ++k) {
       DriftLineRKF drift;
       drift.SetSensor(&sensor);
       drift.SetGainFluctuationsPolya(0., 20000.);
